more types now result is uint32_t

+++ hardware/fpga/f32c/cores/f32c/wiring_shift.c
@@ -24,8 +24,8 @@
 
 #include "wiring_private.h"
 
-uint8_t shiftIn(uint8_t dataPin, uint8_t clockPin, uint8_t bitOrder) {
-	uint8_t value = 0;
+uint32_t shiftIn(uint8_t dataPin, uint8_t clockPin, uint8_t bitOrder) {
+	uint32_t value = 0;
 	uint8_t i;
 
 	for (i = 0; i < 8; ++i) {
